переносит app в папку и подключает в него компоненты
